refactor: 확장 함수를 메서드로 변경
