--- src/main/kotlin/calculator/ParsingService.kt
@@ -1,7 +1,0 @@
-package calculator
-
-fun InputString.toNonNegativeNumbers(): List<NonNegativeNumber> {
-    val delimiter = Delimiter(src)
-    return target.split(delimiter.regex)
-        .map { it.toNonNegativeNumber() }
-}
